Added BatchedTicketRegistryCleaner which supports cleaning of a JpaTicketRegistry in batches to prevent OOM situations
